requirements.txt
Babel==1.3
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.2.11
Flask-Mail==0.7.6
Flask-OpenID==1.2.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.4
Flask-WhooshAlchemy==0.55a
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.7.9
Tempita==0.5.2
WTForms==2.0
Werkzeug==0.9.6
Whoosh==2.6.0
blinker==1.3
decorator==3.4.0
flup==1.0.2
gunicorn==19.0.0
python-openid==2.2.5
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
beautifulsoup4==4.3.2